Introduction - Find

+++ 1.algorithm/3.introduction/2.find/A1011.cpp
@@ -23,13 +23,13 @@ using namespace std;
 int main(int argc, char const *argv[]) {
 
     char ch[3] = {'W', 'T', 'L'};
-    float result = 1;
+    double result = 1;
     for (int i = 0; i < 3; i++) {
-        float o[3];
+        double o[3];
         for (int j = 0; j < 3; j++) {
             cin >> o[j];
         }
-        float *max_ptr = max_element(o, o + 3);
+        double *max_ptr = max_element(o, o + 3);
         result *= *max_ptr;
         cout << ch[max_ptr - o] << " ";
     }
